requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
gradio
pytz